Remove learning tracks feature (#60617)
