+++ content/contributing/writing-for-github-docs/using-yaml-frontmatter.md
@@ -39,7 +39,7 @@ For more information, see [`lib/frontmatter.ts`](https://github.com/github/docs/
 * [`changelog`](#changelog)
 * [`defaultPlatform`](#defaultplatform)
 * [`defaultTool`](#defaulttool)
-* [`learningTracks`](#learningtracks)
+
 * [`journeyTracks`](#journeytracks)
 * [`type`](#type)
 * [`communityRedirect`](#communityredirect)
@@ -225,14 +225,6 @@ defaultPlatform: linux
 defaultTool: cli
 ```
 
-### `learningTracks`
-
-* Purpose: Render a list of learning tracks on a product's sub-landing page.
-* Type: `String`. This should reference learning tracks' names defined in [`data/learning-tracks/*.yml`](https://github.com/github/docs/tree/main/data/learning-tracks).
-* Optional
-
-> [!NOTE]
-> The featured track is set by a specific property in the learning tracks YAML. See that [README](https://github.com/github/docs/blob/main/data/learning-tracks/README.md) for details.
 
 ### `journeyTracks`
 
@@ -320,3 +312,4 @@ Index pages are the table of contents files for the Docs site. Every product, ca
 The homepage is the main Table of Contents file for the docs site. The homepage must have a complete list of `children`, like every [Index page](#index-pages) but must also specify the `childGroups` frontmatter property that will be highlighted in the main content area.
 
 `childGroups` is an array of mappings containing a `name` for the group, an optional `icon` for the group, and an array of `children`. The `children` in the array must be present in the `children` frontmatter property.
+
