+++ package.json
@@ -138,7 +138,6 @@
       "data/variables",
       "data/glossaries",
       "data/product-examples",
-      "data/learning-tracks",
       "rest-api-description",
       "semmle-code"
     ]
