+++ src/data-directory/lib/data-schemas/index.ts
@@ -50,7 +50,6 @@ function loadTableSchemas(): DataSchemas {
 const manualSchemas: DataSchemas = {
   'data/features': resolveSchemaPath('features.ts'),
   'data/variables': resolveSchemaPath('variables.ts'),
-  'data/learning-tracks': resolveSchemaPath('learning-tracks.ts'),
   'data/release-notes': resolveSchemaPath('release-notes.ts'),
   'data/code-languages.yml': resolveSchemaPath('code-languages.ts'),
   'data/glossaries/candidates.yml': resolveSchemaPath('glossaries-candidates.ts'),
