--- src/fixtures/fixtures/data/learning-tracks/code-security.yml
@@ -1,14 +1,0 @@
-foo_bar:
-  title: Fix the plumbing
-  description: A couple easy ones
-  guides:
-    - /code-security/getting-started/quickstart
-    - /code-security/getting-started/securing-your-organization
-
-bar_foo:
-  title: Plumb the fixing
-  description: The reverse of the one above
-  guides:
-    - /code-security/getting-started/securing-your-organization
-    - /code-security/getting-started/quickstart
-    - /rest/actions
