--- src/fixtures/fixtures/data/learning-tracks/codespaces.yml
@@ -1,6 +1,0 @@
-# Codespaces learning tracks
-foo_bar:
-  title: Sample learning track
-  description: A sample track for testing
-  guides:
-    - /get-started/start-your-journey/hello-world
