+++ src/fixtures/fixtures/data/ui.yml
@@ -306,11 +306,6 @@ product_landing:
   view: View all
   all_docs: 'All {{ title }} docs'
   all_content: 'View all {{ title }} content'
-learning_track_nav:
-  prev_guide: Previous
-  next_guide: Next
-  more_guides: More guides →
-  current_progress: '{i} of {n} in learning path'
 journey_track_nav:
   prev_article: Previous
   next_article: Next
