+++ src/frame/components/context/ArticleContext.tsx
@@ -3,16 +3,6 @@ import { createContext, useContext } from 'react'
 import type { JSX } from 'react'
 import type { JourneyContext } from '@/journeys/lib/journey-path-resolver'
 
-export type LearningTrack = {
-  trackTitle: string
-  trackName: string
-  trackProduct: string
-  prevGuide?: { href: string; title: string }
-  nextGuide?: { href: string; title: string }
-  numberOfGuides: number
-  currentGuideIndex: number
-}
-
 export type MiniTocItem = {
   platform?: string
   contents: {
@@ -34,7 +24,6 @@ export type ArticleContextT = {
   defaultPlatform?: string
   defaultTool?: string
   product?: string
-  currentLearningTrack?: LearningTrack
   currentJourneyTrack?: JourneyContext
   detectedPlatforms: Array<string>
   detectedTools: Array<string>
@@ -99,7 +88,6 @@ export const getArticleContextFromRequest = (req: any): ArticleContextT => {
     defaultPlatform: page.defaultPlatform || '',
     defaultTool: page.defaultTool || '',
     product: page.product || '',
-    currentLearningTrack: req.context.currentLearningTrack,
     currentJourneyTrack: req.context.currentJourneyTrack,
     detectedPlatforms: page.detectedPlatforms || [],
     detectedTools: page.detectedTools || [],
