+++ src/journeys/lib/get-link-data.ts
@@ -1,2 +1,1 @@
-// Re-export the getLinkData function from learning tracks for journey tracks
-export { default } from '@/learning-track/lib/get-link-data'
+export { default } from '@/frame/lib/get-link-data'
