+++ src/landings/context/LandingContext.tsx
@@ -2,7 +2,7 @@ import { createContext, useContext } from 'react'
 import { getFeaturedLinksFromReq } from '@/landings/components/ProductLandingContext'
 import { mapRawTocItemToTocItem } from '@/landings/types'
 import type { TocItem } from '@/landings/types'
-import type { ExtendedRequest, Context, LearningTrack } from '@/types'
+import type { ExtendedRequest, Context } from '@/types'
 import type { JourneyTrack } from '@/journeys/lib/journey-path-resolver'
 import type { FeaturedLink } from '@/landings/components/ProductLandingContext'
 
@@ -18,7 +18,6 @@ export type LandingContextT = {
   variant?: 'compact' | 'expanded'
   featuredLinks: Record<string, Array<FeaturedLink>>
   renderedPage: string
-  currentLearningTrack?: LearningTrack | null
   currentLayout: string
   heroImage?: string
   // For landing pages with carousels
@@ -105,7 +104,6 @@ export const getLandingContextFromRequest = async (
     variant: context.genericTocFlat ? 'expanded' : 'compact',
     featuredLinks: getFeaturedLinksFromReq(req),
     renderedPage: context.renderedPage ?? '',
-    currentLearningTrack: context.currentLearningTrack ?? null,
     currentLayout: context.currentLayoutName ?? '',
     heroImage: page.heroImage || '/assets/images/banner-images/hero-1',
     introLinks: page.introLinks || null,
