+++ src/landings/pages/product.tsx
@@ -182,18 +182,12 @@ export const getServerSideProps: GetServerSideProps<Props> = async (context) => 
       props.categoryLandingContext = getCategoryLandingContextFromRequest(req)
     } else {
       props.tocLandingContext = getTocLandingContextFromRequest(req)
-      if (props.tocLandingContext.currentLearningTrack?.trackName) {
-        additionalUINamespaces.push('learning_track_nav')
-      }
     }
   } else if (props.mainContext.page) {
     // All articles that might have hover cards needs this
     additionalUINamespaces.push('popovers')
 
     props.articleContext = getArticleContextFromRequest(req)
-    if (props.articleContext.currentLearningTrack?.trackName) {
-      additionalUINamespaces.push('learning_track_nav')
-    }
     if (props.articleContext.currentJourneyTrack?.trackId) {
       additionalUINamespaces.push('journey_track_nav')
     }
