+++ src/rest/pages/category.tsx
@@ -1,12 +1,7 @@
 import { GetServerSideProps } from 'next'
 import { Operation } from '@/rest/components/types'
 import { RestReferencePage } from '@/rest/components/RestReferencePage'
-import {
-  addUINamespaces,
-  getMainContext,
-  MainContext,
-  MainContextT,
-} from '@/frame/components/context/MainContext'
+import { getMainContext, MainContext, MainContextT } from '@/frame/components/context/MainContext'
 import {
   AutomatedPageContext,
   AutomatedPageContextT,
@@ -209,9 +204,6 @@ export const getServerSideProps: GetServerSideProps<Props> = async (context) => 
   tocLandingContext.tocItems = restCategoryTocItems
 
   const mainContext = await getMainContext(req, res)
-  if (tocLandingContext.currentLearningTrack?.trackName) {
-    addUINamespaces(req, mainContext.data.ui, ['learning_track_nav'])
-  }
 
   return {
     props: {
