+++ src/types/types.ts
@@ -57,7 +57,6 @@ export type PageFrontmatter = {
   featuredLinks?: FeaturedLinks
   changelog?: ChangeLog
   contentType?: string
-  learningTracks?: string[]
   beta_product?: boolean
   interactive?: boolean
   communityRedirect?: {
@@ -178,45 +177,11 @@ export type Context = {
   productUserExamples?: ProductExample[]
   productGroups?: ProductGroup[]
   featuredLinks?: FeaturedLinksExpanded
-  currentLearningTrack?: LearningTrack | null
   renderedPage?: string
   miniTocItems?: MiniTocItem[]
   markdownRequested?: boolean
   markdownViaUrl?: boolean
 }
-export type LearningTracks = {
-  [group: string]: {
-    [track: string]: {
-      title: string
-      description: string
-      versions?: FrontmatterVersions
-      guides: string[]
-    }
-  }
-}
-export type LearningTrack = {
-  trackName: string
-  trackProduct: string
-  trackTitle: string
-  numberOfGuides?: number
-  currentGuideIndex?: number
-  nextGuide?: {
-    href: string
-    title: string | undefined
-  }
-  prevGuide?: {
-    href: string
-    title: string | undefined
-  }
-}
-
-export type TrackGuide = {
-  href: string
-  page: Page
-  title: string
-  intro: string
-}
-
 export type FeaturedLinkExpanded = {
   href: string
   title: string
