+++ src/versions/scripts/use-short-versions.ts
@@ -104,7 +104,7 @@ async function main() {
   for (const file of yamlFiles) {
     const yamlContent = fs.readFileSync(file, 'utf8')
     const yamlReplacements = getLiquidReplacements(yamlContent, file)
-    // Update any `versions` properties in the YAML as well (learning tracks, etc.)
+    // Update any `versions` properties in the YAML as well
     const newYamlContent = makeLiquidReplacements(yamlReplacements, yamlContent)
       .replace(/("|')?free-pro-team("|')?:/g, 'fpt:')
       .replace(/("|')?enterprise-server("|')?:/g, 'ghes:')
